requirements.txt
fastapi
uvicorn
httpx
python-dotenv
vaderSentiment
transformers
torch
tiktoken
protobuf
sentencepiece
pipeline
ninja
cmake
google-api-python-client
trafilatura
beautifulsoup4
lxml
spacy
pytextrank
